move QueryResultsList to stream_generator.py
