+++ google/cloud/firestore_v1/query.py
@@ -29,7 +29,6 @@ from google.cloud import firestore_v1
 from google.cloud.firestore_v1 import aggregation, transaction
 from google.cloud.firestore_v1.base_document import (
     DocumentSnapshot,
-    QueryResultsList,
 )
 from google.cloud.firestore_v1.base_query import (
     BaseCollectionGroup,
@@ -39,7 +38,7 @@ from google.cloud.firestore_v1.base_query import (
     _enum_from_direction,
     _query_response_to_snapshot,
 )
-from google.cloud.firestore_v1.stream_generator import StreamGenerator
+from google.cloud.firestore_v1.stream_generator import QueryResultsList, StreamGenerator
 from google.cloud.firestore_v1.vector import Vector
 from google.cloud.firestore_v1.vector_query import VectorQuery
 from google.cloud.firestore_v1.watch import Watch
